Split core feature gate type tests
