+++ src/typechecker/tests/core_semantics/feature_gates/gated_types.rs
@@ -1,0 +1,178 @@
+use super::*;
+
+#[test]
+fn typed_allocator_type_is_rejected_as_gated_not_unknown() {
+    let program = parse_program(
+        r#"
+main = (allocator: Allocator<i32, Sync>) void { }
+"#,
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program(&program)
+        .expect_err("typed allocator types should stay gated until allocator semantics exist");
+
+    assert!(
+        err.iter()
+            .any(|d| d.message.contains("typed allocators are gated")),
+        "expected typed allocator gate diagnostic, got {err:?}"
+    );
+    assert!(
+        err.iter()
+            .all(|d| !d.message.contains("unknown type symbol")),
+        "allocator/effect gate should not be reported as ordinary unknown types, got {err:?}"
+    );
+}
+
+#[test]
+fn sync_and_async_typed_allocator_modes_are_rejected_as_gated_not_unknown() {
+    let program = parse_program(
+        r#"
+use_sync_allocator = (allocator: Allocator<i32, Sync>) void { }
+use_async_allocator = (allocator: Allocator<i32, Async>) void { }
+"#,
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc.check_program(&program).expect_err(
+        "typed allocator Sync/Async modes should stay gated until allocator semantics exist",
+    );
+
+    let allocator_gate_count = err
+        .iter()
+        .filter(|diagnostic| diagnostic.message.contains("typed allocators are gated"))
+        .count();
+    assert_eq!(
+        allocator_gate_count, 2,
+        "expected Sync and Async allocator annotations to both report allocator gates, got {err:?}"
+    );
+    assert!(
+        err.iter()
+            .all(|d| !d.message.contains("unknown type symbol")),
+        "allocator/effect mode gates should not be reported as ordinary unknown types, got {err:?}"
+    );
+}
+
+#[test]
+fn dynamic_string_type_is_rejected_as_allocator_backed_gate() {
+    let program = parse_program(
+        r#"
+main = (value: String) void { }
+"#,
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program(&program)
+        .expect_err("dynamic String should stay gated until allocator ownership exists");
+
+    assert!(
+        err.iter().any(|d| d.message.contains("`String` is gated")),
+        "expected dynamic String gate diagnostic, got {err:?}"
+    );
+    assert!(
+        err.iter()
+            .all(|d| !d.message.contains("unknown type symbol")),
+        "dynamic String gate should not be reported as an ordinary unknown type, got {err:?}"
+    );
+}
+
+#[test]
+fn sync_async_effect_modes_are_rejected_as_gated_not_unknown() {
+    let program = parse_program(
+        r#"
+use_sync = (mode: Sync) void { }
+use_async = (mode: Async) void { }
+"#,
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program(&program)
+        .expect_err("effect mode types should stay gated until effect semantics exist");
+
+    for expected in [
+        "`Sync` effect mode is gated",
+        "`Async` effect mode is gated",
+    ] {
+        assert!(
+            err.iter()
+                .any(|diagnostic| diagnostic.message.contains(expected)),
+            "expected diagnostic `{expected}`, got {err:?}"
+        );
+    }
+    assert!(
+        err.iter()
+            .all(|d| !d.message.contains("unknown type symbol")),
+        "effect mode gate should not be reported as ordinary unknown types, got {err:?}"
+    );
+}
+
+#[test]
+fn actor_framework_types_are_rejected_as_gated_not_unknown() {
+    let program = parse_program(
+        r#"
+use_actor = (actor: Actor<i32>) void { }
+use_ref = (ref: ActorRef<i32>) void { }
+use_mailbox = (mailbox: Mailbox<i32>) void { }
+use_supervisor = (supervisor: Supervisor) void { }
+"#,
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program(&program)
+        .expect_err("actor framework types should stay gated until std actor semantics exist");
+
+    for expected in [
+        "`Actor` framework type is gated",
+        "`ActorRef` framework type is gated",
+        "`Mailbox` framework type is gated",
+        "`Supervisor` framework type is gated",
+    ] {
+        assert!(
+            err.iter()
+                .any(|diagnostic| diagnostic.message.contains(expected)),
+            "expected diagnostic `{expected}`, got {err:?}"
+        );
+    }
+    assert!(
+        err.iter()
+            .all(|d| !d.message.contains("unknown type symbol")),
+        "actor framework gate should not be reported as ordinary unknown types, got {err:?}"
+    );
+}
+
+#[test]
+fn bare_actor_framework_types_are_rejected_as_gated_not_unknown() {
+    let program = parse_program(
+        r#"
+use_actor = (actor: Actor) void { }
+use_ref = (ref: ActorRef) void { }
+use_mailbox = (mailbox: Mailbox) void { }
+"#,
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program(&program)
+        .expect_err("bare actor framework types should stay gated until std actor semantics exist");
+
+    for expected in [
+        "`Actor` framework type is gated",
+        "`ActorRef` framework type is gated",
+        "`Mailbox` framework type is gated",
+    ] {
+        assert!(
+            err.iter()
+                .any(|diagnostic| diagnostic.message.contains(expected)),
+            "expected diagnostic `{expected}`, got {err:?}"
+        );
+    }
+    assert!(
+        err.iter()
+            .all(|d| !d.message.contains("unknown type symbol")),
+        "bare actor framework gates should not be reported as ordinary unknown types, got {err:?}"
+    );
+}
