+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -1,3 +1,4 @@
+mod core_semantics_splits;
 mod focused_tests;
 mod function_method_template_splits;
 mod lexer_and_monomorphize;
